// apps/blog/lib/images/generated.ts
// This file is auto-generated by scripts/generate-images.ts. Do not edit manually.

export const imagesManifest = {
  "language-models-can-use-existing-software": {
    "cover": {
      "width": 1347,
      "height": 386,
      "placeholder": "data:image/webp;base64,UklGRmwAAABXRUJQVlA4IGAAAADQAwCdASoUAAYAPrVInkmnJCKhMAgA4BaJaQDG9CPs3Ful6V0kIcAA/umdAmgJErmIzjlZcvT1jPiK2VrmR8MMoqQPcHFWyjsUO4m7hMToCVMhfyL4R6DqpT1bGv8AAAA=",
      "variants": [
        {
          "width": 480,
          "height": 138,
          "url": "/blog-assets/language-models-can-use-existing-software/cover-480.webp"
        },
        {
          "width": 768,
          "height": 220,
          "url": "/blog-assets/language-models-can-use-existing-software/cover-768.webp"
        },
        {
          "width": 1024,
          "height": 293,
          "url": "/blog-assets/language-models-can-use-existing-software/cover-1024.webp"
        },
        {
          "width": 1280,
          "height": 367,
          "url": "/blog-assets/language-models-can-use-existing-software/cover-1280.webp"
        }
      ]
    },
    "opengraph-image": {
      "width": 1347,
      "height": 386,
      "placeholder": "data:image/webp;base64,UklGRmYAAABXRUJQVlA4IFoAAACwAwCdASoUAAYAPrVInkmnJCKhMAgA4BaJaQDDNBeDlML/huVtIAD+5QJMKHcTRoT3BVRd/imE1tuk0JyDlJnB17OHYfit7c9xEjeSJIM1ao731r74WJggAAA=",
      "variants": [
        {
          "width": 480,
          "height": 138,
          "url": "/blog-assets/language-models-can-use-existing-software/opengraph-image-480.webp"
        },
        {
          "width": 768,
          "height": 220,
          "url": "/blog-assets/language-models-can-use-existing-software/opengraph-image-768.webp"
        },
        {
          "width": 1024,
          "height": 293,
          "url": "/blog-assets/language-models-can-use-existing-software/opengraph-image-1024.webp"
        },
        {
          "width": 1280,
          "height": 367,
          "url": "/blog-assets/language-models-can-use-existing-software/opengraph-image-1280.webp"
        }
      ]
    },
    "platform-spectrum": {
      "width": 1189,
      "height": 372,
      "placeholder": "data:image/webp;base64,UklGRkwAAABXRUJQVlA4IEAAAADQAwCdASoUAAYAPrVKnkmnJCKhMAgA4BaJZwDE2CKV/sKTI/KR8AAA/vLYbhSNXIvBtYVZB1ghNmU9FCynNiAA",
      "variants": [
        {
          "width": 480,
          "height": 150,
          "url": "/blog-assets/language-models-can-use-existing-software/platform-spectrum-480.webp"
        },
        {
          "width": 768,
          "height": 240,
          "url": "/blog-assets/language-models-can-use-existing-software/platform-spectrum-768.webp"
        },
        {
          "width": 1024,
          "height": 320,
          "url": "/blog-assets/language-models-can-use-existing-software/platform-spectrum-1024.webp"
        }
      ]
    }
  }
} as const;